requirements.txt
fastapi
uvicorn[standard]
Hydro-Quebec-API-Wrapper
python-dotenv
